requirements.txt
fastapi>=0.115.9
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic>=2.5.0
alembic==1.12.0
python-dotenv==1.0.0
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-chroma>=0.2.4
chromadb>=1.0.9
python-multipart>=0.0.6 